Fix compilation bug with TBB v2018.0
Previously (e.g. in TBB v4.4) a tbb::combinable could be initialized by
an explicit constant. However in later versions (e.g. TBB v2018.0) the
initializer must be a callable function. As a result we now use an
explicit function to perform zero initialization for the tbb::combinable's
which track uninitialized nodes in the timing graph.