Add support for applications to report detailed edge delay breakdowns
Host applications can now override the TimingGraphNameResolver::
edge_delay_breakdown() to return a set of named delay components
which make up the aggregate delay of an edge.

This can be useful, for example in FPGA tools where timing graph edges
are typically a combination of different device components, to give
users some addition insight into how their designs are implemented,
and why the delays are what they are.