Fix increment offset in timing reports
Since we now don't annotate the clock period constraint on CLOCK_LAUNCH
tags during the traversal (since it may be capture node dependent due to
multicycles), we need to account for the offset when reporting the path.
This ensures the increments line up correctly with the calculated path
required times.